refactor: 카테고리 별 이미지 컴포넌트로 분리
